Add ./configure --enable-fontforge-extras
Bug 827962 - acorn2sfd not in fontforge package.
Very old bug, but only saw this recently.
acorn2sfd now builds, plus added acorn2sfd.1 man page.
Moved acorn2sfd.c out of the libfontforge directory since it was an
orphan program (most users will want just fontforge), and moved it
into fonttools where it now can be part of fontforge-extras.
Interestingly ARM stands for Acorn RISC Machine.

Bug 754414 fix - showttf.c: "Vender URL" should be "Vendor URL"
merged patch shown in launchpad by Brian Murray <brian@ubuntu.com>

Bug #736221 - showttf -verbose prints help message "showttf -verbose"
instead of being verbose
Reference to a man page? found one, but also needed extra options added.
Enabled building showttf to see what was the problem with this bug here.
Wasn't able to reproduce bug, but further checking saw this bug was
already patched by Alan Coopersmith (merged by Michal-n 2010-04-15).

import/sync "fontforge-extras" showttf.c and this showttf.c code_of
interest to update this version. 'diff' showed some missed edits or
additions worth adding. Then did a lot of editing to reduce the
large number of warnings and errors showing during compiling.

Add missing Encodings.ps.gz (found in fontforge-extras).
put it in doc/html since it is referenced in html docs.
TODO: fix FontForge to get this file in native scripting.

Updated contrib files so they compile using option:
./configure --enable-fontforge-extras.
Default is currently no install. Can be changed later.

Updated stripttc with error checks.
converted README-ttf2eps into ttf2eps.1

Fixed configure.ac to pay more attention to glib-2.0 pkg
and flags. This also involved check headers for -pthreads
because GLIB_CFLAGS outputs pthreads and GLIB_LIBS outputs
several libraries. There also seems to be another type of
GIO-2.0 called GIO-UNIX-2.0 distributed by some distros.

PThreads checking needs to be careful of which platform it
is on, building on linux is easy, but some of the other
host platforms expect other libraries, therefore kept the
checks minimal to just lib function and pthreads.h check.

cidmap/src programs are now built (but not installed).

Modified travis.yml to avoid the make && make contrib.

+++ contrib/cidmap/src/mapkorean.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include "utype.h"
 
 static char used[0x110000];
 static int cid_2_unicode[0x10000];
@@ -87,7 +88,7 @@ return( -1 );
 		mults[j++] = vals[i];
 	}
 	if ( j>MULT_MAX ) {
-	    fprintf( stderr, "Too many multiple values for %04x, need %d slots\n", val, j );
+	    fprintf( stderr, "Too many multiple values for %04x, need %d slots\n", (unsigned int)(val), j );
 exit(1);
 	}
     }
@@ -122,7 +123,7 @@ int main(int argc, char **argv) {
 	    /*if ( psunicodenames[uni]!=NULL )
 		sprintf( buffer, "%s.hw", psunicodenames[uni]);
 	    else*/
-		sprintf( buffer, "uni%04X.hw", fakeuni);
+		sprintf( buffer, "uni%04X.hw", (unsigned int)(fakeuni) );
 	    nonuni_names[cid] = strdup(buffer);
 	} else if ( cid>=18255 && cid<=18351 ) {
 	    /* rotated halfwidth latin */
@@ -143,7 +144,7 @@ int main(int argc, char **argv) {
 	    used[uni] = 1;
 	    cid_2_unicode[cid] = uni;
 	} else {
-	    sprintf( buffer, "uni%04X.dup%d", uni, ++used[uni] );
+	    sprintf( buffer, "uni%04X.dup%d", (unsigned int)(uni), ++used[uni] );
 	    nonuni_names[cid] = strdup(buffer);
 	}
 	max = cid;
@@ -153,7 +154,7 @@ int main(int argc, char **argv) {
 	    if ( cid_2_unicode[j] == cid_2_rotunicode[i] )
 	break;
 	if ( j==maxcid )
-	    sprintf( buffer, "uni%04X.vert", cid_2_rotunicode[i]);
+	    sprintf( buffer, "uni%04X.vert", (unsigned int)(cid_2_rotunicode[i]) );
 	else
 	    sprintf( buffer, "Korea1.%d.vert", j);
 	nonuni_names[i] = strdup(buffer);
@@ -166,14 +167,14 @@ int main(int argc, char **argv) {
 	    for ( i=1; cid+i<=max && cid_2_unicode[cid+i]==cid_2_unicode[cid]+i && cid_2_unicodemult[cid+i][0]==0; ++i );
 	    --i;
 	    if ( i!=0 ) {
-		printf( "%d..%d %04x\n", cid, cid+i, cid_2_unicode[cid] );
+		printf( "%d..%d %04x\n", cid, cid+i, (unsigned int)(cid_2_unicode[cid]) );
 		cid += i;
 	    } else
-		printf( "%d %04x\n", cid, cid_2_unicode[cid] );
+		printf( "%d %04x\n", cid, (unsigned int)(cid_2_unicode[cid]) );
 	} else if ( cid_2_unicode[cid]!=-1 ) {
-	    printf( "%d %04x", cid, cid_2_unicode[cid]);
+	    printf( "%d %04x", cid, (unsigned int)(cid_2_unicode[cid]) );
 	    for ( i=0; cid_2_unicodemult[cid][i]!=0; ++i )
-		printf( ",%04x", cid_2_unicodemult[cid][i]);
+		printf( ",%04x", (unsigned int)(cid_2_unicodemult[cid][i]) );
 	    printf( "\n");
 	} else if ( nonuni_names[cid]!=NULL )
 	    printf( "%d /%s\n", cid, nonuni_names[cid] );
